third_party/blink/web_tests/fast/dom/shadow/caret-range-from-point-in-nested-shadow-expected.txt

PASS internals.treeScopeRootNode(range.startContainer) is document
PASS internals.treeScopeRootNode(range.endContainer) is document
PASS successfullyParsed is true

TEST COMPLETE
